refactor: usePostDetail 커스텀 훅 작성
